### 2025-09-15 #8922 @Kevin-Mattheus-Moerman
@cfrioux Dear author, thanks for this submission. I am the AEiC on this track and here to help process the initial steps.

Please note that JOSS is currently managing a large backlog of submissions and the editors most appropriate for your area are all already rather busy.
