### 2025-12-15 975395f
enhance ontology management for cpd exploration
