+++ m2m_postaviz/cpd_exploration_module.py
@@ -14,13 +14,63 @@ from m2m_postaviz.data_struct import DataStorage
 @module.ui
 def cpd_tab_ui(Data: DataStorage):
 
+    # Build first-level options: top-level categories first, then all others (prefixed)
+    def _first_selector_options():
+        tree = Data.get_cpd_category_tree()
+        # Unwrap artificial root
+        unwrapped = tree
+        if len(unwrapped) == 1 and all(isinstance(v, dict) for v in unwrapped.values()):
+            (_, unwrapped) = next(iter(unwrapped.items()))
+
+        # Top-level options (direct children)
+        data_cpd_list = Data.get_compound_list(without_compartment=True)
+        top_level = []
+        for key, child in unwrapped.items():
+            if not child:
+                continue
+            cpd_in_category = []
+            Data.get_compounds_from_category(child, cpd_in_category)
+            final_cpd_list = [cpd for cpd in data_cpd_list if cpd in cpd_in_category]
+            top_level.append(f"{key} ({len(final_cpd_list)}/{len(cpd_in_category)})")
+
+        # All categories (entire tree)
+        all_with_counts = Data.get_metacyc_category_list()
+
+        # Merge with top-level first, then the rest (excluding duplicates and the single artificial root if present)
+        seen = set()
+        merged = []
+        for item in top_level:
+            merged.append(item)
+            seen.add(item.split(" (")[0])
+        for item in all_with_counts:
+            name = item.split(" (")[0]
+            if name in ("All_metabolites",):
+                continue
+            if name not in seen:
+                merged.append(f"-- {item}")  # prefix to visually separate deeper nodes
+                seen.add(name)
+        return merged
+
     if Data.USE_METACYC_PADMET:
 
-        cpd_exploration_all_category = Data.get_metacyc_category_list()
+        cpd_exploration_all_category = _first_selector_options()
 
-        cpd_exploration_all_category.insert(0, Data.get_outsider_cpd()[1])
+        # Add global All/Others with counts for the first selector
+        data_cpd_list = Data.get_compound_list(without_compartment=True)
+        all_cpd_list = []
+        Data.get_compounds_from_category(Data.get_cpd_category_tree(), all_cpd_list)
+        outsiders_full = Data.get_outsider_cpd()[0]
+        all_cpd_list = list(set(all_cpd_list + outsiders_full))
+        all_in_data = len([cpd for cpd in all_cpd_list if cpd in data_cpd_list])
+        all_label = f"All ({all_in_data}/{len(all_cpd_list)})"
 
-        cpd_exploration_all_category.insert(0,"None")
+        outsiders = outsiders_full
+        outsiders_in_data = len([cpd for cpd in outsiders if cpd in data_cpd_list])
+        outsiders_label = f"Others ({outsiders_in_data}/{len(all_cpd_list)})"
+
+        cpd_exploration_all_category.insert(0, outsiders_label)
+        cpd_exploration_all_category.insert(0, all_label)
+        cpd_exploration_all_category.insert(0, "None")
 
     else:
 
@@ -36,8 +86,10 @@ def cpd_tab_ui(Data: DataStorage):
         ui.layout_sidebar(
             ui.sidebar(
 
-                ui.input_selectize("first_category_input","Select any compound category from the Metacyc database.\n(Compounds in data / Compounds in metacyc database)",choices=cpd_exploration_all_category, multiple=False),
-                ui.input_selectize("category_choice_input", "Select a sub category of the Metacyc database", choices=cpd_exploration_all_category, selected=cpd_exploration_all_category[0], multiple=False, width="400px"),
+                ui.input_selectize("first_category_input","Select any compound category from the Metacyc database (optional).\n(Compounds in data / Compounds in metacyc database)",choices=cpd_exploration_all_category, multiple=False),
[…]
