### 2026-01-12 bec8e5d
fix GA for macos-latest
