+++ paper/paper.md
@@ -24,23 +24,29 @@ bibliography: paper.bib
 
 # Summary
 
-Microbial communities consist of up to thousands of distinct microbial populations, each characterized by its genomic DNA sequences, and all sharing a habitat and environmental conditions. The word _microbiome_ describe the holistic concept associated all the previous components [@Berg.2020]. Characterizing the populations in samples and understanding both their roles of and interactions within the microbiome require a combination of experimentation, high-throughput data acquisition and computational models [@Klitgord.2011]. The role of microbial populations can be abstracted by the study of their metabolism, encompassing all biochemical reactions that cells may perform. The genome encodes the genetic information associated to the metabolism, and can therefore be queried to estimate the metabolic capabilities of microorganisms, represented as metabolic networks. Because such metabolism is highly redundant, depends on the environment, and because of the size of microbial communities, computational models are needed and predict, from the networks, the possible behaviours of and interactions within microbial populations in given environmental conditions. Comparing the outcomes of such predictive models in multiple samples further increases the difficulty of integrating results into actionable hypotheses. Visualisation and integration of metadata can help perform such comparison. They need to be as customisable as possible to facilitate exploration by end-users.  
+Microbial communities consist of up to thousands of distinct microbial populations, each characterized by its genomic DNA sequences, and all sharing a habitat and environmental conditions. The word _microbiome_ describe the holistic concept associated all the previous components [@Berg.2020]. Characterising the populations in samples and understanding both their roles of and interactions within the microbiome require a combination of experimentation, high-throughput data acquisition and computational models [@Klitgord.2011]. The role of microbial populations can be abstracted by the study of their _metabolism_, encompassing all biochemical reactions that cells may perform. The genome encodes the genetic information associated to the metabolism, and can therefore be queried to estimate the metabolic capabilities of microorganisms, represented as _metabolic networks_. Because such metabolism is highly redundant, depends on the environment, and because of the size of microbial communities, computational models are needed and predict, from the networks, the possible behaviours of and interactions within microbial populations in given environmental conditions. Comparing the outcomes of such predictive models in multiple samples further increases the difficulty of integrating results into actionable hypotheses. Visualisation and integration of metadata can help perform such comparison. They need to be as customisable as possible to facilitate exploration by end-users.  
 
 # Statement of need
-
-Metabolic modelling is widely used to determine the roles of microorganisms in microbial communities [@Cerk.2024]: what molecules they can produce, which interactions are likely to happen depending on the environmental conditions... Many models rely on integer linear programming optimisations [@García-Jimenez.2021] and some alternatives use Boolean abstractions [@Frioux.2018] [@Belcour.2020] or probabilistic approaches [@Bernstein.2019] to provide predictions on community behaviours. Key questions are scalability to large communities, and also integration of these predictions when analysing many samples or community compositions. Each microorganism is abstracted by a collection of biochemical reactions it may perform according to its genomic information, thus forming a network connecting transformed molecules. A community of microorganisms is therefore represented as a collection of such networks, referred to as genome-scale metabolic networks. A metabolic model will provide predictions on the possible behaviour of microorganisms and communities in defined simulation conditions.
 
[…]
