+++ paper/paper.bib
@@ -31,6 +31,20 @@
 }
 
 
+@article{Asnicar.2025,
+  year      = {2025},
+  title     = {{Gut micro-organisms associated with health, nutrition and dietary interventions}},
+  author    = {Asnicar, Francesco and Manghi, Paolo and Fackelmann, Gloria and Baldanzi, Gabriel and Bakker, Elco and Ricci, Liviana and Piccinno, Gianmarco and Piperni, Elisa and Mladenovic, Katarina and Amati, Federica and Arrè, Alberto and Ganesh, Sajaysurya and Giordano, Francesca and Davies, Richard and Wolf, Jonathan and Bermingham, Kate M. and Berry, Sarah E. and Spector, Tim D. and Segata, Nicola},
+  journal   = {Nature},
+  issn      = {0028-0836},
+  doi       = {10.1038/s41586-025-09854-7},
+  pmid      = {41372407},
+  abstract  = {{The incidence of cardiometabolic diseases is increasing globally, and both poor diet and the human gut microbiome have been implicated1. However, the field lacks large-scale, comprehensive studies exploring these links in diverse populations2. Here, in over 34,000 US and UK participants with metagenomic, diet, anthropometric and host health data, we identified known and yet-to-be-cultured gut microbiome species associated significantly with different diets and risk factors. We developed a ranking of species most favourably and unfavourably associated with human health markers, called the ‘ZOE Microbiome Health Ranking 2025’. This system showed strong and reproducible associations between the ranking of microbial species and both body mass index and host disease conditions on more than 7,800 additional public samples. In an additional 746 people from two dietary interventional clinical trials, favourably ranked species increased in abundance and prevalence, and unfavourably ranked species reduced over time. In conclusion, these analyses provide strong support for the association of both diet and microbiome with health markers, and the summary system can be used to inform the basis for future causal and mechanistic studies. It should be emphasized, however, that causal inference is not possible without prospective cohort studies and interventional clinical trials.}},
+  pages     = {1--9},
+  keywords  = {},
+  local-url = {file://localhost/Users/cfrioux/Documents/Papers%20Library/Asnicar-Gut%20micro-organisms%20associated%20with%20health,%20nutrition%20and%20dietary%20interventions-2025-Nature_1.pdf}
+}
+
 @article{Belcour.2020,
     year      = {2020},
     title     = {{Metage2Metabo, microbiota-scale metabolic complementarity for the identification of key species}},
