+++ paper/paper.bib
@@ -31,6 +31,20 @@
 }
 
 
+@article{Kruse.2008,
+  year      = {2008},
+  title     = {{Comparing flux balance analysis to network expansion: producibility, sustainability and the scope of compounds.}},
+  author    = {Kruse, Kai and Ebenhöh, Oliver},
+  journal   = {Genome informatics. International Conference on Genome Informatics},
+  issn      = {0919-9454},
+  pmid      = {19425125},
+  abstract  = {{The producibility of metabolites from available resources is investigated systematically using flux balance analysis (FBA) and network expansion. Calculations are performed for the genome-scale metabolic networks of Escherichia coli and Methanosarcina barkeri. Strict biological interpretation of the results obtained with FBA leads to the concept of sustainability, which reduces the set of producible metabolites by assuming a growing and dividing cell. A systematic comparison showed that applying network expansion in many cases results in exactly the set of all sustainable metabolites. The purely heuristic approach of allowing for certain cofactors to facilitate reactions during the process of network expansion dramatically helps to improve agreement of the results from the two different approaches. In conclusion, we state that network expansion, due to its enormous advantages in computational speed, is a valuable alternative to determining producible metabolites with FBA.}},
+  pages     = {91--101},
+  volume    = {20},
+  keywords  = {},
+  local-url = {file://localhost/Users/cfrioux/Documents/Papers%20Library/Kruse-Comparing%20flux%20balance%20analysis%20to%20network%20expansion-%20producibility,%20sustainability%20and%20the%20scope%20of%20compounds--2008-Genome%20informatics.%20International%20Conference%20on%20Genome%20Informatics_1.pdf}
+}
+
 @article{Asnicar.2025,
   year      = {2025},
   title     = {{Gut micro-organisms associated with health, nutrition and dietary interventions}},
