### 2026-02-06 bdcdffb
suspect latest python-libsbml leads to segfault in GA
