+++ requirements.txt
@@ -5,9 +5,10 @@ plotly>=5.18.0
 shiny>=0.6.0
 shinywidgets>=0.2.3
 scikit-bio>=0.5.9
+pyarrow
+python-libsbml>=5.20.5,<5.21.0
 seaborn>=0.13.2
 polars>=1.12.0
 polars-lts-cpu>=0.15.9
 sphinx_rtd_theme>=3.0.2
 faicons>=0.2.2
-
