### 2026-02-06 83a129a
latest pyarrow version suspected to break tests
