### 2026-02-06 c359d46
latest pyarrow version suspected to break tests
