+++ pyproject.toml
@@ -18,7 +18,7 @@ dependencies = [
   'shiny',
   'shinywidgets',
   'scikit-bio',
-  'pyarrow>=14.0.0,<23.0.0',
+  'pyarrow==22.0.0',
   'python-libsbml>=5.20.5,<5.21.0',
   'seaborn',
 ]
