### 2026-02-06 3466d87
pyarrow workaround to be tested on ubuntu
