### 2026-02-15 #8922 @arfon
@cfrioux – thanks for updating the paper. Your Research Impact Statement talks about the impact of the M2M project, but has this software itself been used for any research results (either by you or community researchers)? If so, please could you cite those outputs?
