### 2026-02-17 #8922 @cfrioux
M2M-postaviz has been used for research results but none have been published yet. So my idea for this statement was to show that M2M is already used, hence a tool that facilitates the interpretation of many M2M runs could be impactful as well.

Regarding the related JOSS papers, maybe the ones with applications to microbiome data (TAXPASTA, Microbiome.jl) are more relevant than the others.
