### 2026-03-16 #8922 @Kevin-Mattheus-Moerman
@cfrioux if other works using this software can be pushed to a pre-print stage it would be great to cite them in the current JOSS paper as evidence the work is being used. Can you look into this?
