### 2026-03-16 #8922 @cfrioux
@Kevin-Mattheus-Moerman No, there is no preprint available unfortunately.
