### 2026-03-17 #8922 @erik-whiting
Hi @cfrioux, thanks for submitting to JOSS, I'm really excited to be the editor for your submission. I am going to try to find some reviewers based on availability and workload, once I have two agreed to review, I will start the formal review process. You are welcome to suggest reviewers if you know any, but I will begin by searching through our list of volunteer reviewers.
