### 2026-03-17 #8922 @erik-whiting
Hello @rastanton, I see you have volunteered to review for JOSS. It looks like your expertise lines up nicely with this submission, would you be willing to review?
