### 2026-03-17 #8922 @erik-whiting
Hello @rpetit3, you have reviewed a couple of submissions for JOSS before, would you be willing to lend your time and expertise to this submission as well?
