### 2026-03-17 #8922 @rastanton
@erik-whiting Sure, I'd be happy to review.

Thanks--Rich

> Hello [@rastanton](https://github.com/rastanton), I see you have volunteered to review for JOSS. It looks like your expertise lines up nicely with this submission, would you be willing to review?
