### 2026-04-09 #8922 @erik-whiting
Hello @abhishektiwari, you've recently expressed interest in reviewing for JOSS. It looks like your expertise lines up nicely with this work, would you be willing to review this submission for us?
