### 2026-06-05 #10650 @editorialbot
[REVIEW]: Metage2Metabo PostAViz: a Python package for exploring, visualising, and comparing the metabolic potential of microbial communities
**Submitting author:** @cfrioux (<a href="http://orcid.org/0000-0003-2114-0697">Clémence Frioux</a>)
**Repository:** https://github.com/AuReMe/metage2metabo-postaviz
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.1
