### 2026-06-26 770eb2b
test python 3.14, add badges, add missing dep
