+++ requirements.txt
@@ -1,7 +1,9 @@
 pandas>=1.5.3
 scipy==1.14.1
+statsmodels>=0.14.0
 padmet>=5.0.1
 plotly>=5.18.0
+matplotlib>=3.8.0
 shiny>=0.6.0
 shinywidgets>=0.2.3
 scikit-bio>=0.5.9
@@ -9,6 +11,5 @@ pyarrow==22.0.0
 python-libsbml>=5.20.5,<5.21.0
 seaborn>=0.13.2
 polars>=1.12.0
-polars-lts-cpu>=0.15.9
 sphinx_rtd_theme>=3.0.2
 faicons>=0.2.2
