+++ README.md
@@ -1,4 +1,4 @@
-[![PyPI version](https://img.shields.io/pypi/v/m2m-postaviz.svg)](https://pypi.org/project/m2m-postaviz/) [![GitHub license](https://img.shields.io/github/license/AuReMe/metage2metabo-postaviz.svg)](https://github.com/AuReMe/metage2metabo-postaviz/blob/main/LICENSE) [![Actions Status](https://github.com/AuReMe/metage2metabo-postaviz/actions/workflows/pythonpackage.yml/badge.svg)](https://github.com/AuReMe/metage2metabo-postaviz/actions/workflows/pythonpackage.yml) [![Documentation Status](https://readthedocs.org/projects/metage2metabo-postaviz/badge/?version=latest)](https://metage2metabo-postaviz.readthedocs.io/en/latest/?badge=latest)
+[![PyPI version](https://img.shields.io/pypi/v/m2m-postaviz.svg)](https://pypi.org/project/m2m-postaviz/) [![GitHub license](https://img.shields.io/github/license/AuReMe/metage2metabo-postaviz.svg)](https://github.com/AuReMe/metage2metabo-postaviz/blob/main/LICENSE) [![Build 3.12](https://img.shields.io/github/check-runs/AuReMe/metage2metabo-postaviz/main?checkName=build-3.12)](https://github.com/AuReMe/metage2metabo-postaviz/actions/workflows/pythonpackage.yml) [![Build 3.13](https://img.shields.io/github/check-runs/AuReMe/metage2metabo-postaviz/main?checkName=build-3.13)](https://github.com/AuReMe/metage2metabo-postaviz/actions/workflows/pythonpackage.yml) [![Build 3.14](https://img.shields.io/github/check-runs/AuReMe/metage2metabo-postaviz/main?checkName=build-3.14)](https://github.com/AuReMe/metage2metabo-postaviz/actions/workflows/pythonpackage.yml) [![Documentation Status](https://readthedocs.org/projects/metage2metabo-postaviz/badge/?version=latest)](https://metage2metabo-postaviz.readthedocs.io/en/latest/?badge=latest)
 
 # Metage2Metabo-PostAViz (M2M-PostAViz)
 
@@ -6,7 +6,7 @@ M2M-PostAViz (_M2M Post-Analysis and Visualization_) is an interactive platform 
 
 ## Installation
 
-The application is tested on Python versions 3.12 and 3.13, on Windows, MacOS and Ubuntu.
+The application is tested on Python versions 3.12, 3.13 and 3.14, on Windows, MacOS and Ubuntu.
 
 Install with pip:
 
