+++ .github/workflows/pythonpackage.yml
@@ -22,30 +22,66 @@ on:
       - '.readthedocs.yml'
 
 jobs:
-  build:
+  build-3.12:
     strategy:
       matrix:
-        platform: [ubuntu-latest, windows-latest] #, macos-latest 
-        python-version: [3.12, 3.13]
+        platform: [ubuntu-latest, windows-latest] #, macos-latest
     runs-on: ${{ matrix.platform }}
 
     steps:
     - uses: actions/checkout@v2
-    - name: Set up Python ${{ matrix.python-version }}
+    - name: Set up Python 3.12
       uses: actions/setup-python@v2
       with:
-        python-version: ${{ matrix.python-version }}
+        python-version: '3.12'
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
         pip install -r requirements.txt
-    #- name: Lint with flake8
-      #run: |
-        #pip install flake8
-        # stop the build if there are Python syntax errors or undefined names
-        #flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-        # exit-zero treats all errors as warnings. The GitHub editor is 127 chars wide
-        #flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+    - name: Test with pytest
+      run: |
+        pip install pytest
+        pip install .
+        pytest
+
+  build-3.13:
+    strategy:
+      matrix:
+        platform: [ubuntu-latest, windows-latest] #, macos-latest
+    runs-on: ${{ matrix.platform }}
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python 3.13
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.13'
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Test with pytest
+      run: |
+        pip install pytest
+        pip install .
+        pytest
+
+  build-3.14:
+    strategy:
+      matrix:
+        platform: [ubuntu-latest, windows-latest] #, macos-latest
+    runs-on: ${{ matrix.platform }}
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python 3.14
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.14'
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
     - name: Test with pytest
       run: |
         pip install pytest
