+++ pyproject.toml
@@ -13,18 +13,21 @@ requires-python = ">= 3.8"
 dependencies = [
   'pandas',
   'scipy',
+  'statsmodels',
   'padmet',
   'plotly',
+  'matplotlib',
   'shiny',
   'shinywidgets',
+  'faicons',
   'scikit-bio',
+  'polars>=1.12.0',
   'pyarrow==22.0.0',
   'python-libsbml>=5.20.5,<5.21.0',
   'seaborn',
 ]
 
 [project.optional-dependencies]
-cpu = ["polars"]
 lts-cpu = ["polars-lts-cpu"]
 
 [project.scripts]
