### 2026-06-26 3b81141
document fallback for polar failure
