### 2026-06-26 92d1792
tentative fix for badges and GH CI
