+++ .github/workflows/pythonpackage.yml
@@ -22,7 +22,7 @@ on:
       - '.readthedocs.yml'
 
 jobs:
-  build-3.12:
+  build_3_12:
     strategy:
       matrix:
         platform: [ubuntu-latest, windows-latest] #, macos-latest
@@ -44,7 +44,7 @@ jobs:
         pip install .
         pytest
 
-  build-3.13:
+  build_3_13:
     strategy:
       matrix:
         platform: [ubuntu-latest, windows-latest] #, macos-latest
@@ -66,7 +66,7 @@ jobs:
         pip install .
         pytest
 
-  build-3.14:
+  build_3_14:
     strategy:
       matrix:
         platform: [ubuntu-latest, windows-latest] #, macos-latest
