### 2026-06-26 b98c4f1
fix dep for py 3.14
