+++ requirements.txt
@@ -1,5 +1,5 @@
 pandas>=1.5.3
-scipy==1.14.1
+scipy>=1.15.0
 statsmodels>=0.14.0
 padmet>=5.0.1
 plotly>=5.18.0
