+++ pyproject.toml
@@ -12,7 +12,7 @@ license = {text = "LGPL-3.0-or-later"}
 requires-python = ">= 3.8"
 dependencies = [
     'pandas',
-    'scipy',
+    'scipy>=1.15.0',
     'statsmodels',
     'padmet',
     'plotly',
