### 2026-06-26 a812fc7
fix the support of Python 3.14 CI and avoid import-time padmet crashes
