+++ paper/paper.bib
@@ -1,4 +1,134 @@
+@manual{Arrow.2026,
+    title  = {arrow: Integration to 'Apache' 'Arrow'},
+    author = {Neal Richardson and Ian Cook and Nic Crane and Dewey Dunnington and Romain François and Jonathan Keane and Bryce Mecum and Dragoș Moldovan-Grünfeld and Jeroen Ooms and Jacob Wujciak-Jens and {Apache Arrow}},
+    year   = {2026},
+    note   = {R package version 24.0.0},
+    url    = {https://github.com/apache/arrow/}
+}
 
+@software{reback2020pandas,
+    author    = {The pandas development team},
+    title     = {pandas-dev/pandas: Pandas},
+    month     = feb,
+    year      = 2020,
+    publisher = {Zenodo},
+    version   = {latest},
+    doi       = {10.5281/zenodo.3509134},
+    url       = {https://doi.org/10.5281/zenodo.3509134}
+}
+
+@article{Waskom2021,
+    doi       = {10.21105/joss.03021},
+    url       = {https://doi.org/10.21105/joss.03021},
+    year      = {2021},
+    publisher = {The Open Journal},
+    volume    = {6},
+    number    = {60},
+    pages     = {3021},
+    author    = {Michael L. Waskom},
+    title     = {seaborn: statistical data visualization},
+    journal   = {Journal of Open Source Software}
+}
+
+@inproceedings{mckinney-proc-scipy-2010,
+    author    = { {W}es {M}c{K}inney },
+    title     = { {D}ata {S}tructures for {S}tatistical {C}omputing in {P}ython },
+    booktitle = { {P}roceedings of the 9th {P}ython in {S}cience {C}onference },
+    pages     = { 56 - 61 },
+    year      = { 2010 },
+    editor    = { {S}t\'efan van der {W}alt and {J}arrod {M}illman },
+    doi       = { 10.25080/Majora-92bf1922-00a }
+}
+
+@article{Hunter:2007,
+    author    = {Hunter, J. D.},
+    title     = {Matplotlib: A 2D graphics environment},
+    journal   = {Computing in Science \& Engineering},
+    volume    = {9},
+    number    = {3},
+    pages     = {90--95},
+    abstract  = {Matplotlib is a 2D graphics package used for Python for
+                application development, interactive scripting, and publication-quality
+                image generation across user interfaces and operating systems.},
+    publisher = {IEEE COMPUTER SOC},
+    doi       = {10.1109/MCSE.2007.55},
+    year      = 2007
+}
+
+@software{polars,
+    author = {{The Polars development team}},
+    title  = {Polars: Extremely fast Query Engine for DataFrames, written in Rust},
+    year   = {2024},
+    url    = {https://pola.rs},
+    note   = {https://github.com/pola-rs/polars}
+}
+
+@software{shiny,
+    author = {{The Shiny development team}},
+    title  = {Shiny for Python},
+    year   = {2024},
+    url    = {https://shiny.posit.co/py/},
+    note   = {https://github.com/posit-dev/py-shiny}
+}
+
+@inproceedings{seabold2010statsmodels,
+    title     = {statsmodels: Econometric and statistical modeling with python},
+    author    = {Seabold, Skipper and Perktold, Josef},
+    booktitle = {9th Python in Science Conference},
+    year      = {2010}
+}
+
+@article{Aite2018,
+    year     = {2018},
+    title    = {Traceability, reproducibility and wiki-exploration for “à-la-carte” reconstructions of genome-scale metabolic models},
[…]
