+++ paper/paper.md
@@ -1,5 +1,5 @@
 ---
-title: 'Metage2Metabo PostaViz: a Python package for exploring, visualising, and comparing the metabolic potential of microbial communities'
+title: 'Metage2Metabo PostAViz: a Python package for exploring, visualising, and comparing the metabolic potential of microbial communities'
 tags:
   - Python
   - Systems biology
@@ -24,19 +24,28 @@ bibliography: paper.bib
 
 # Summary
 
-Microbial communities consist of up to thousands of distinct microbial populations, each characterized by its genomic DNA sequences, and all sharing a habitat and environmental conditions. The word _microbiome_ describe the holistic concept associated all the previous components [@Berg.2020]. Characterising the populations in samples and understanding both their roles of and interactions within the microbiome require a combination of experimentation, high-throughput data acquisition and computational models [@Klitgord.2011]. The role of microbial populations can be abstracted by the study of their _metabolism_, encompassing all biochemical reactions that cells may perform. The genome encodes the genetic information associated to the metabolism, and can therefore be queried to estimate the metabolic capabilities of microorganisms, represented as _metabolic networks_. Because such metabolism is highly redundant, depends on the environment, and because of the size of microbial communities, computational models are needed and predict, from the networks, the possible behaviours of and interactions within microbial populations in given environmental conditions. Comparing the outcomes of such predictive models in multiple samples further increases the difficulty of integrating results into actionable hypotheses. Visualisation and integration of metadata can help perform such comparison. They need to be as customisable as possible to facilitate exploration by end-users.  
+Microbial communities comprise up to several thousand distinct microbial populations, each defined by its genomic DNA sequences and sharing a common habitat and environmental conditions. The term _microbiome_ describes the holistic concept encompassing these microorganisms and their environment [@Berg.2020]. Characterising the populations within samples and understanding their roles and interactions in the microbiome requires a combination of experimentation, high-throughput data acquisition and computational modelling [@Klitgord.2011].
+
+The role of microbial populations can be abstracted through the study of their _metabolism_, encompassing the full set of biochemical reactions performed by cells. Within a microbiome, metabolic functions often display both redundancy and complementarity: distinct microbial compositions may support similar overall functions. This functional overlap motivates the systematic inference of metabolic potential. Because the genome encodes the information underlying metabolism, it can be used to estimate the metabolic capabilities of microorganisms, typically represented as _metabolic networks_. These networks form the basis of _metabolic models_, which integrate additional data within mathematical or computational frameworks to predict the behaviour of microbial communities under specific environmental conditions.
+
+Advances in experimental strategies and large-scale DNA sequencing in metagenomics now enable the analysis of increasing numbers of samples. Scalable approaches are therefore required to explore potential behaviours and interactions among microbial populations across environments. Comparing predictions across multiple samples is essential for generating actionable hypotheses, but doing so requires integration of sample metadata, statistical analysis, and careful interpretation of modelling outputs. Effective visualisation and customisable exploration tools are therefore important to make metabolic modelling results accessible and usable for end users.  
+
 
 # Statement of need
 
[…]
