### 2026-06-26 83adc17
fix badges and min py version in paper
