+++ paper/paper.md
@@ -53,7 +53,7 @@ In practice, each microorganism is abstracted by a collection of biochemical rea
 
 ## Data integration and storage
 
-Each sample must first be processed with `M2M`. The inputs to `M2M-PostAViz` are the directories generated by `M2M` for each sample, which contain multiple tabulated and JSON files. Additional inputs include microbial relative abundances (used to weight predictions), sample metadata (e.g. lifestyle or clinical variables), and optionally metadata describing metabolites and microbial taxonomy. These inputs are provided as tabular files.
+`M2M-PostaViz` is implemented in Python (≥3.12). Each sample must first be processed with `M2M`. The inputs to `M2M-PostAViz` are the directories generated by `M2M` for each sample, which contain multiple tabulated and JSON files. Additional inputs include microbial relative abundances (used to weight predictions), sample metadata (e.g. lifestyle or clinical variables), and optionally metadata describing metabolites and microbial taxonomy. These inputs are provided as tabular files.
 
 Integrating these heterogeneous data sources across hundreds or thousands of samples can be computationally demanding, particularly since each sample corresponds to a separate output directory. Moreover, users typically explore the same dataset across multiple sessions. To avoid repeated parsing and integration of raw outputs, M2M-PostAViz implements a preprocessing step that consolidates and structures all required data into a unified representation.
 
