### 2026-06-30 79f9f9e
change CLI to support interactive mode
