+++ m2m_postaviz/cli.py
@@ -15,13 +15,21 @@ Why does this file exist, and why not put this in __main__?
   Also see (1) from http://click.pocoo.org/5/setuptools/#setuptools-integration
 """
 import argparse
+import sys
 import tempfile
+from typing import Optional
 
 from m2m_postaviz import __version__ as VERSION
 import m2m_postaviz.data_utils as du
-import m2m_postaviz.shiny_app as sh
 from pathlib import Path
 from m2m_postaviz.data_struct import DataStorage
+
+try:
+    from prompt_toolkit import prompt as pt_prompt
+    from prompt_toolkit.completion import PathCompleter
+except ModuleNotFoundError:
+    pt_prompt = None
+    PathCompleter = None
 
 LICENSE = """Copyright (C) L. brindel and C. Frioux.\n
 This program is free software: you can redistribute it and/or modify
@@ -63,75 +71,210 @@ TESTS_DIR = Path(SRC_DIR, "postaviz_test_data/")
 
 data_table_filepath = Path(TESTS_DIR, "table_test_postaviz.tar.gz")
 
+
+class InteractivePromptExit(Exception):
+    """Raised when user exits interactive prompt."""
+
+
+def _is_interactive_terminal() -> bool:
+    return sys.stdin.isatty() and sys.stdout.isatty()
+
+
+def _check_exit_command(answer: str) -> None:
+    """Raise InteractivePromptExit if user typed exit command."""
+    if answer.lower() in {"quit", "exit", "q"}:
+        raise InteractivePromptExit("User exited interactive prompt.")
+
+
+def _read_path_input(label: str, only_directories: bool = False) -> str:
+    # Use shell-like path completion when prompt_toolkit is available.
+    hint = " (or 'quit'/'exit' to cancel)"
+    full_label = label + hint
+    if pt_prompt is not None and PathCompleter is not None:
+        completer = PathCompleter(only_directories=only_directories, expanduser=True)
+        return pt_prompt(f"{full_label}: ", completer=completer).strip()
+    return input(f"{full_label}: ").strip()
+
+
+def _prompt_yes_no(label: str, default: bool = False) -> bool:
+    suffix = " [Y/n]" if default else " [y/N]"
+    hint = " (or 'quit'/'exit' to cancel)"
+    answer = input(f"{label}{suffix}{hint}: ").strip()
+    _check_exit_command(answer)
+    answer_lower = answer.lower()
+    if not answer_lower:
+        return default
+    return answer_lower in {"y", "yes"}
+
+
+def _prompt_required_path(label: str, must_exist: bool = True, is_dir: bool = False) -> Path:
+    while True:
+        answer = _read_path_input(label, only_directories=is_dir)
+        _check_exit_command(answer)
+        if not answer:
+            print("Value cannot be empty.")
+            continue
+
+        path = Path(answer).resolve()
+
+        if not must_exist:
+            return path
+
+        if not path.exists():
+            print(f"Path does not exist: {path}")
+            continue
+
+        if is_dir and not path.is_dir():
+            print(f"Expected a directory: {path}")
+            continue
+
+        if not is_dir and not path.is_file():
+            print(f"Expected a file: {path}")
+            continue
+
+        return path
+
+
+def _prompt_optional_path(label: str, expect_file: bool = True) -> Optional[Path]:
+    while True:
+        answer = _read_path_input(label, only_directories=not expect_file)
+        _check_exit_command(answer)
+        if not answer or answer.lower() in {"skip", "none"}:
+            return None
+
+        path = Path(answer).resolve()
+
+        if not path.exists():
+            print(f"Path does not exist: {path}. Press Enter to skip.")
+            continue
+
+        if expect_file and not path.is_file():
[…]
