+++ tests/cli_test.py
@@ -1,0 +1,194 @@
+from pathlib import Path
+
+from m2m_postaviz import cli
+
+
+def test_main_non_interactive_missing_required(monkeypatch, capsys):
+    called = {"build": False}
+
+    def fake_build_dataframes(*_args, **_kwargs):
+        called["build"] = True
+
+    monkeypatch.setattr(cli, "_is_interactive_terminal", lambda: False)
+    monkeypatch.setattr(cli.du, "build_dataframes", fake_build_dataframes)
+
+    cli.main([])
+
+    out = capsys.readouterr().out
+    assert "Required arguments missing" in out
+    assert called["build"] is False
+
+
+def test_main_interactive_precomputed_bypass(monkeypatch, tmp_path):
+    load_dir = tmp_path / "precomputed"
+    load_dir.mkdir()
+
+    captured = {"load_path": None, "ran_shiny": False}
+
+    class DummyDataStorage:
+        def __init__(self, path):
+            captured["load_path"] = Path(path)
+
+    def fake_run_shiny(_data):
+        captured["ran_shiny"] = True
+
+    monkeypatch.setattr(cli, "_is_interactive_terminal", lambda: True)
+    monkeypatch.setattr(cli, "_prompt_yes_no", lambda *args, **kwargs: True)
+    monkeypatch.setattr(cli, "_prompt_required_path", lambda *args, **kwargs: load_dir)
+    monkeypatch.setattr(cli, "DataStorage", DummyDataStorage)
+    monkeypatch.setattr(cli, "_run_shiny", fake_run_shiny)
+
+    cli.main([])
+
+    assert captured["load_path"] == load_dir
+    assert captured["ran_shiny"] is True
+
+
+def test_main_interactive_required_and_optional_skip(monkeypatch, tmp_path):
+    data_dir = tmp_path / "m2m_data"
+    data_dir.mkdir()
+
+    metadata_file = tmp_path / "metadata.tsv"
+    metadata_file.write_text("sample\tgroup\nS1\tA\n", encoding="utf-8")
+
+    output_dir = tmp_path / "out"
+
+    captured = {
+        "dir_path": None,
+        "metadata_path": None,
+        "abundance_path": None,
+        "taxonomy_path": None,
+        "save_path": None,
+        "metacyc_path": None,
+        "ran_shiny": False,
+        "data_path": None,
+    }
+
+    def fake_build_dataframes(*args):
+        captured["dir_path"] = args[0]
+        captured["metadata_path"] = args[1]
+        captured["abundance_path"] = args[2]
+        captured["taxonomy_path"] = args[3]
+        captured["save_path"] = args[4]
+        captured["metacyc_path"] = args[5]
+
+    class DummyDataStorage:
+        def __init__(self, path):
+            captured["data_path"] = Path(path)
+
+    def fake_run_shiny(_data):
+        captured["ran_shiny"] = True
+
+    required_values = iter([data_dir, metadata_file, output_dir])
+
+    monkeypatch.setattr(cli, "_is_interactive_terminal", lambda: True)
+    monkeypatch.setattr(cli, "_prompt_yes_no", lambda *args, **kwargs: False)
+    monkeypatch.setattr(cli, "_prompt_required_path", lambda *args, **kwargs: next(required_values))
+    monkeypatch.setattr(cli, "_prompt_optional_path", lambda *args, **kwargs: None)
+
+    monkeypatch.setattr(cli.du, "build_dataframes", fake_build_dataframes)
+    monkeypatch.setattr(cli, "DataStorage", DummyDataStorage)
+    monkeypatch.setattr(cli, "_run_shiny", fake_run_shiny)
+
+    cli.main([])
+
+    assert captured["dir_path"] == data_dir
[…]
