### 2026-06-30 0415a88
clarify CLI information
