### 2026-06-30 d7b271d
include macos in GH actions
