+++ .github/workflows/pythonpackage.yml
@@ -25,7 +25,7 @@ jobs:
   build_3_12:
     strategy:
       matrix:
-        platform: [ubuntu-latest, windows-latest] #, macos-latest
+        platform: [ubuntu-latest, windows-latest, macos-latest] #
     runs-on: ${{ matrix.platform }}
 
     steps:
@@ -47,7 +47,7 @@ jobs:
   build_3_13:
     strategy:
       matrix:
-        platform: [ubuntu-latest, windows-latest] #, macos-latest
+        platform: [ubuntu-latest, windows-latest, macos-latest] #
     runs-on: ${{ matrix.platform }}
 
     steps:
@@ -69,7 +69,7 @@ jobs:
   build_3_14:
     strategy:
       matrix:
-        platform: [ubuntu-latest, windows-latest] #, macos-latest
+        platform: [ubuntu-latest, windows-latest, macos-latest] #
     runs-on: ${{ matrix.platform }}
 
     steps:
