+++ paper/paper.md
@@ -47,7 +47,7 @@ Overall, `M2M-PostAViz` was designed to address three complementary needs: (i) e
 
 Metabolic modelling is widely used to determine the roles of microorganisms in microbial communities [@Cerk.2024]: what molecules they can produce, which interactions are likely to happen depending on the environmental conditions... Many models rely on integer linear programming optimisations [@García-Jimenez.2021] and some alternatives use Boolean abstractions [@Frioux.2018] [@Belcour.2020] or probabilistic approaches [@Bernstein.2019] to provide predictions on community behaviours. Key questions are scalability to large communities, and also integration of these predictions when analysing many samples or community compositions, which is more and more frequent as large metagenomic cohorts get published [@Asnicar.2025].
 
-In practice, each microorganism is abstracted by a collection of biochemical reactions it may perform according to its genomic information, thus forming a network connecting transformed molecules [@Cerk.2024]. A community of microorganisms is therefore represented as a collection of such networks, referred to as genome-scale metabolic networks. A metabolic model will provide predictions on the possible behaviour of microorganisms and communities in defined simulation conditions. Boolean abstractions, as used by M2M, ensure the scalability of predictions; although the model is qualitative and does not quantify interactions, such methods are considered a reasonable proxy for more quantitative models [@Kruse.2008].
+In practice, each microorganism is abstracted by a collection of biochemical reactions it may perform according to its genomic information, thus forming a network connecting transformed molecules [@Cerk.2024]. A community of microorganisms is therefore represented as a collection of such networks, referred to as genome-scale metabolic networks. A metabolic model will provide predictions on the possible behaviour of microorganisms and communities in defined simulation conditions. Boolean abstractions, as used by M2M, ensure the scalability of predictions; although the model is qualitative and does not quantify interactions, such methods are considered a reasonable proxy for more quantitative models [@Kruse.2008] (Figure \autoref{fig:postaviz}A).
 
 # Software design
 
@@ -63,7 +63,7 @@ Data manipulation and storage rely primarily on pandas [@reback2020pandas] [@mck
 
 ## Application content
 
-The application opens as a multi-tab browser page. The first tab provides an overview of the dataset and enables initial analyses that can be customised using metadata variables. Two additional tabs provide analyses centred on microbial roles and metabolites, respectively (see below), and a final tab summarises the metadata and allows users to customise variable types to fine-tune analyses and plots. Statistical analyses are performed using scipy [@2020SciPy-NMeth] and statsmodels [@seabold2010statsmodels]. Visualisations are generated with plotly [@plotly], matplotlib [@Hunter:2007], and seaborn [@Waskom2021], and the web interface is implemented using shiny for Python [@shiny].
+The application opens as a multi-tab browser page. The first tab provides an overview of the dataset and enables initial analyses that can be customised using metadata variables. Two additional tabs provide analyses centred on microbial roles and metabolites, respectively (see below), and a final tab summarises the metadata and allows users to customise variable types to fine-tune analyses and plots. Statistical analyses are performed using scipy [@2020SciPy-NMeth] and statsmodels [@seabold2010statsmodels]. Visualisations are generated with plotly [@plotly], matplotlib [@Hunter:2007], and seaborn [@Waskom2021], and the web interface is implemented using shiny for Python [@shiny] (Figure \autoref{fig:postaviz}B).
 
 ### Exploration of microorganism roles
 
@@ -72,6 +72,11 @@ The second tab of the application focuses on the role of microorganisms in the p
 ### Exploration of metabolite production across samples
 
[…]
