### 2026-06-30 ec670e9
remove dummy fig placeholder
