### 2026-06-30 e02b517
fix figure code in paper
