+++ paper/paper.md
@@ -73,7 +73,7 @@ The second tab of the application focuses on the role of microorganisms in the p
 
 The third tab of the application is dedicated to the analysis of metabolites, which can also be grouped into families if a suitable ontology is provided [@Caspi2019] [@Aite2018]. Focusing on metabolites enables a targeted analysis in order to compare samples or groups of samples on specific metabolic functions.
 
-![**Figure 1:** _Illustration of the context and functionalities of Metage2Metabo-PostAViz._ **A.** Metagenomic context. Samples consist of microbial communities, each composed of several interacting microbial populations. These populations are represented as metabolic networks used to predict community behaviour and interactions within dedicated modelling frameworks. **B.** Functionalities of M2M-PostAViz. Inputs include Metage2Metabo outputs for each sample, metadata, relative abundances, and taxonomic information associated with the samples. The application enables comparison of the metabolic potential of samples and exploration of their associations with metadata through a graphical user interface.\label{fig:postaviz}](postaviz_figure.pdf)
+![_Illustration of the context and functionalities of Metage2Metabo-PostAViz._ **A.** Metagenomic context. Samples consist of microbial communities, each composed of several interacting microbial populations. These populations are represented as metabolic networks used to predict community behaviour and interactions within dedicated modelling frameworks. **B.** Functionalities of M2M-PostAViz. Inputs include Metage2Metabo outputs for each sample, metadata, relative abundances, and taxonomic information associated with the samples. The application enables comparison of the metabolic potential of samples and exploration of their associations with metadata through a graphical user interface.\label{fig:postaviz}](postaviz_figure.pdf)
 
 # Research impact statement
 
