### 2026-06-30 51bcb06
CRON scheduling of CI
