+++ docs/conf.py
@@ -18,7 +18,7 @@ project = "m2m-postaviz"
 year = "2023"
 author = "Léonard Brindel, Clémence Frioux"
 copyright = f"{year}, {author}"
-version = release = "0.1.1"
+version = release = "0.2.0"
 
 pygments_style = "trac"
 templates_path = ["."]
