### 2026-06-30 afa628c
troubleshooting install in doc
