### 2026-07-01 b38abb4
test CI GA on python latest
