+++ .github/workflows/pythonpackage.yml
@@ -89,3 +89,25 @@ jobs:
         pip install pytest
         pip install .
         pytest
+
+  build_3_x:
+    strategy:
+      matrix:
+        platform: [ubuntu-latest, windows-latest, macos-latest] #
+    runs-on: ${{ matrix.platform }}
+
+    steps:
+    - uses: actions/checkout@v2
+    - name: Set up Python 3.x
+      uses: actions/setup-python@v2
+      with:
+        python-version: '3.x'
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        pip install -r requirements.txt
+    - name: Test with pytest
+      run: |
+        pip install pytest
+        pip install .
+        pytest
