### 2026-07-09 #10650 @erik-whiting
@cfrioux I think I agree with @benyamindsmith on the GUI as I think it is uncommon to be able to start a GUI without input data. For example, pymol (a tool for visualizing PDB files) can be launched from the command line with an input `.pdb` or it can be started without any input and the user can load their PDB file from the GUI. I think this is more standard than requiring input at launch.

Having said that, I do understand that most use-cases for this software are likely on HPC clusters which usually don't have GUI support. That makes the case that this software is still a useful contribution without the GUI improvements. Personally, I'm leaning more towards considering the GUI request required for acceptance but hopefully @rastanton can take a look soon and weigh in on this discussion as well.

For the record:

>That being said I would prefer to not be the main decision-maker around this feature if my request is not welcome by you

Nothing is unwelcome, I'm very grateful for your time and attention to this review.
