### 2026-08-31 #10650 @benyamindsmith
I didn't mean remove me!!

On Sun, Aug 30, 2026, 9:03 p.m. The Open Journals editorial robot <
***@***.***> wrote:

> *editorialbot* left a comment (openjournals/joss-reviews#10650)
> <https://github.com/openjournals/joss-reviews/issues/10650#issuecomment-5472435297>
>
> @benyamindsmith <https://github.com/benyamindsmith> removed from the
> reviewers list!
>
