### 2026-08-31 #10650 @erik-whiting
Hello @NehaAr, are you willing/able to review this submission? This request is a little outside of protocol for JOSS but we just lost a reviewer mid-review, we'd be super grateful if you could take a look!
