### 2026-08-31 #10650 @erik-whiting
Happy Sunday @98ketaki, I see you've volunteered to review for JOSS. I'm hoping you can join this review that's already in-progress, we need a bioinformatician's perspective I think!
