### 2026-08-31 #10650 @erik-whiting
@NehaAr thank you so much! You can generate your checklist by typing this into the comment box:

In addition to those checks, we are also wondering what you think about item two in [this GitHub issue](https://github.com/AuReMe/metage2metabo-postaviz/issues/6) raised by another reviewer? Basically we are trying to decide if the limitation of not being able to start the GUI part of this software in a "blank" mode without input data. Please let us know what you think, and thank you again!
