### 2026-08-31 #10650 @erik-whiting
@98ketaki thank we appreciate it so much! If you type this command you can generate your reviewer checklist:

I said this above but I'll write it again just in case: In addition to the checklist items, we are wondering what you think about item two in [this GitHub issue](https://github.com/AuReMe/metage2metabo-postaviz/issues/6) raised by another reviewer? Basically we are trying to decide if the limitation of not being able to start the GUI part of this software in a "blank" mode without input data needs to be addressed before we can accept. Please let us know what you think, and thank you again!
