### 2026-09-01 AuReMe/metage2metabo-postaviz#7 @NehaAr
Clarify data-sharing checklist item: is bundled compound classification database 'original data'
Data sharing: If the paper contains original data, data are accessible to the reviewers. If the paper contains no original data, please check this item.

If this compound classification file in the data folder is something you (or your package) built, curated, or substantially modified—e.g., you compiled it from MetaCyc/other sources, added your own structure, or it's your own resource.
