### 2026-09-02 #10650 @erik-whiting
> [@erik-whiting](https://github.com/erik-whiting)
>
> [@cfrioux](https://github.com/cfrioux) I have opened the following issue for my review:
>
> [Clarify data-sharing checklist item: is bundled compound classification database 'original data' #7](https://github.com/AuReMe/metage2metabo-postaviz/issues/7)
>
> Please address these issues when you can.
>
> Also , I noticed that some in-text citations do not appear to follow the expected Pandoc citation syntax. For example, citations are sometimes written as [@Frioux.2018] [@Belcour.2020] or [[@Polars](https://github.com/Polars)]. Could the authors please check the citation syntax throughout the manuscript and ensure that all references are correctly formatted and rendered according to the JOSS/Pandoc citation format?
>
> and I noticed a few instances where three consecutive dots (...) are used in the manuscript in State of the field

amazing, thank you so much for this review!

Do you have any thoughts on starting the GUI discussion? Do you think the app should be able to be open without input data to be accepted?
