### 2026-09-03 #10650 @NehaAr
@erik-whiting
I agree that a blank-start GUI could improve usability. While large datasets may be better handled through the CLI/HPC, this does not necessarily prevent the GUI from starting blank and allowing users to select their data. Could this approach be considered or clarified in the documentation if not feasible?
