### 2026-09-05 #10650 @NehaAr
@cfrioux

Also noticed some unused files in the repo, like authors. rst in  docs, and readme. rst ... Those could be deleted
